pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "iartisanxl"
version = "0.0.1"
authors = [
  { name="Alvaro Somoza", email="somoza.alvaro@gmail.com" },
]
description = "Dekstop application for generating images using Stable Diffusion."
requires-python = ">=3.11"
dependencies = [
    "accelerate>=0.24.1 ",
    "diffusers>=0.22.0",
    "Pillow>=9.3.0",
    "PyOpenGL",
    "PyOpenGL_accelerate",
    "PyQt6>=6.6.0",
    "pyqtgraph",
    "safetensors",
    "scipy",
    "torchsde",
    "transformers>=4.35.0",
    "toml",
    "tomesd",
    "attrs",
    "omegaconf",
    "PyOpenGL",
    "peft>=0.6.0",    
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-qt",
    "pytest-cov",
    "build",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
iartisanxl = ["theme/stylesheet.qss", "theme/icons/*.*", "theme/images/*.*"]

[project.gui-scripts]
iartisanxl = "iartisanxl.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--import-mode=importlib",
]
filterwarnings = [
    "ignore::DeprecationWarning:torchvision.transforms._functional_pil",
]